SL-1767: rename listing to hearing
