+++ src/app/app.module.ts
@@ -39,7 +39,7 @@ import { AuthHttpInterceptor } from './security/services/auth-http-interceptor';
 import { SecurityContext } from './security/services/security-context.service';
 import { HmctsModule } from './hmcts/hmcts.module';
 import { GovukModule } from './govuk/govuk.module';
-import { ListingModule } from './listing/listing.module';
+import { HearingModule } from './hearing/hearing.module';
 
 @Injectable()
 export class XhrInterceptor implements HttpInterceptor {
@@ -89,7 +89,7 @@ export class XhrInterceptor implements HttpInterceptor {
         TransactionsModule,
         HmctsModule,
         GovukModule,
-        ListingModule
+        HearingModule
     ],
     providers: [SessionsService, AppConfig, AppConfigGuard, SecurityService, SecurityContext,
         {provide: HTTP_INTERCEPTORS, useClass: XhrInterceptor, multi: true},
