+++ src/app/hearing/components/view-hearing/view-hearing.component.scss
